trying to run the auction test, thereis somthing up with concurrency and no one seems to buy anything

+++ src/main/scala/InitialPractice/model/company/asset/Asset.scala
@@ -1,5 +1,5 @@
 package InitialPractice.model.company.asset
 
-class Asset (var value:BigDecimal, val id:Int) {
+class Asset (var value:BigDecimal, val id:Long) {
 
 }
